Move format_lines into printing module
